Flask_Blog/blog-src/requirements.txt
alembic==1.4.3
autopep8==1.5.4
bcrypt==3.2.0
blinker==1.4
cachelib==0.1.1
cffi==1.14.3
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
Pillow==7.2.0
pycodestyle==2.6.0
pycparser==2.20
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
uWSGI==2.0.19.1
Werkzeug==1.0.1
WTForms==2.3.3
